# swagger.yaml
swagger: "2.0"
info:
  description: "API manager with CTS Sales"
  version: 1.0.0
  title: API Manager
basePath: /v1
schemes:
  - http
  - https
paths:
...
  "/service/sobjects":
    get:
      tags:
        - service
      summary: Get sobject
      description: Returns Object
      operationId: sobjects
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
        - application/xml
      responses:
        "200":
          description: Successful operation
        "404":
          description: Object not found
      security:
        - Bearer: []
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
...